requirements.txt
flask
flask-socketio
eventlet
opencv-python
numpy
PyYAML
rpi_ws281x
adafruit-circuitpython-neopixel
